Make coverage tracker local to function invocation
Previously, we used v8::internal::coverage_tracker as a
global variable for convenience of passing the coverage
tracker along. However, this is a significant limitation
if we wish to multi-thread fuzzing in the future.

The CoverageTracker object is now passed all the way to
the interpreter via function parameters to avoid
any need for shared memory coordination.